+++ libdraw/arith.c
@@ -159,7 +159,7 @@ combinerect(Rectangle *r1, Rectangle r2)
 		r1->max.y = r2.max.y;
 }
 
-ulong drawld2chan[] = {
+uint32 drawld2chan[] = {
 	GREY1,
 	GREY2,
 	GREY4,
@@ -171,8 +171,8 @@ ulong drawld2chan[] = {
  * to avoid updating many binaries on sources.
  */
 
-ulong
-setalpha(ulong color, uchar alpha)
+uint32
+setalpha(uint32 color, uchar alpha)
 {
 	int red, green, blue;
