+++ libmemlayer/lalloc.c
@@ -5,7 +5,7 @@
 #include "memlayer.h"
 
 Memimage*
-memlalloc(Memscreen *s, Rectangle screenr, Refreshfn refreshfn, void *refreshptr, ulong val)
+memlalloc(Memscreen *s, Rectangle screenr, Refreshfn refreshfn, void *refreshptr, uint32 val)
 {
 	Memlayer *l;
 	Memimage *n;
